Refactor z80 plugin to not have global state ##arch
